AMBARI-10043. Test Kerberos Client (KERBEROS_SERVICE_CHECK) is failed after Ambari only upgrade from 1.6.0 to 2.0.0 when enable securitty (Centos 5.9) (rlevas)

+++ ambari-server/src/main/resources/common-services/KERBEROS/1.10.3-10/package/scripts/service_check.py
@@ -52,7 +52,8 @@ class KerberosServiceCheck(KerberosScript):
         # kinit
         Execute(kinit_command)
       finally:
-        os.remove(ccache_file_path)
+        if os.path.isfile(ccache_file_path): # Since kinit might fail to write to the cache file for various reasons, an existence check should be done before cleanup
+          os.remove(ccache_file_path)
     else:
       err_msg = Logger.filter_text("Failed to execute kinit test due to principal or keytab not found or available")
       raise Fail(err_msg)
